+++ src/redux/playerSlice.js
@@ -1,0 +1,8 @@
+import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
+import userApi from "./userApi";
+
+const playersSlice = createSlice({
+   name: "player"
+})
+
+export default playerSlice.reducer;
